+++ tests/test_named_tuple.py
@@ -135,11 +135,6 @@ class TestPositive:
         for key, val in g:
             assert d[key] == val
 
-    def test_len(self, init_namedtuple) -> None:
-        """Get the number of elements."""
-        nt = init_namedtuple
-        assert len(nt) == 2
-
     def test_keys(self, init_namedtuple) -> None:
         """Get a list of keys."""
         nt = init_namedtuple
@@ -176,3 +171,28 @@ class TestPositive:
         x = nt["x"]
         x = 5
         assert nt["x"] != x
+
+    def test_repr(self, init_namedtuple) -> None:
+        """Test a `__repr__` method."""
+        nt = init_namedtuple
+
+        assert repr(nt) == "NamedTuple(x=10, y={})".format('"Hello"')
+
+    def test_str(self, init_namedtuple) -> None:
+        """Test a `__str__` method."""
+        nt = init_namedtuple
+
+        assert str(nt) == str({"x": 10, "y": "Hello"})
+
+    def test_bool(self, init_namedtuple) -> None:
+        """Test a `__bool__` method."""
+        nt = init_namedtuple
+        assert bool(nt)
+
+        nt = NamedTuple()
+        assert not bool(nt)
+
+    def test_len(self, init_namedtuple) -> None:
+        """Get the number of elements."""
+        nt = init_namedtuple
+        assert len(nt) == 2